+++ test/entities/ratingTest.js
@@ -28,7 +28,7 @@ var CourseSection = require('../../src/entities/agent/courseSection');
 var DigitalResource = require('../../src/entities/resource/digitalResource');
 var DigitalResourceCollection = require('../../src/entities/resource/digitalResourceCollection');
 var LikertScale = require('../../src/entities/survey/likertScale');
-var RatingScaleQuestion = require('../../src/entities/survey/RatingScaleQuestion');
+var RatingScaleQuestion = require('../../src/entities/question/RatingScaleQuestion');
 var Rating = require('../../src/entities/survey/rating');
 var clientUtils = require('../../src/clients/clientUtils');
 var testUtils = require('../testUtils');
