+++ test/events/feedbackEventRankedTest.js
@@ -34,7 +34,7 @@ var LikertScale = require('../../src/entities/survey/likertScale');
 var Membership = require('../../src/entities/agent/membership');
 var Person = require('../../src/entities/agent/person');
 var Rating = require('../../src/entities/survey/rating');
-var RatingScaleQuestion = require('../../src/entities/survey/RatingScaleQuestion');
+var RatingScaleQuestion = require('../../src/entities/question/RatingScaleQuestion');
 var Role = require('../../src/entities/agent/role');
 var Session = require('../../src/entities/session/session');
 var SoftwareApplication = require('../../src/entities/agent/softwareApplication');
